Put the aws lambda functions to scripts. Refine some implementation.
