--- python/mspasspy/cloud/aws_lambda_func_def.py
@@ -1,10 +1,0 @@
-from obspy.core.stream import Stream
-from obspy.core import UTCDateTime
-
-def lambda_func(input_path, event):
-    # after processing, return output_path
-    # arg1 = event['arg1']
-    # arg2 = event['arg2']
-    # output_path = preprocess(intput_path)
-    # return output_path
-    pass
