--- python/mspasspy/cloud/aws_settings.py
@@ -1,9 +1,0 @@
-# Settings for aws lambda uploading and calling.
-
-aws_access_key_id = ''
-
-aws_secret_access_key = ''
-
-lambda_upload_bucket = ''
-
-lambda_iam_role = ''
